Streamline test setup functions with parameterize
