+++ tests/test_transformer/test_object_transformer_new.py
@@ -1,15 +1,9 @@
 """Test the object transformer. -- New framework"""
-import pytest
 from linkml_runtime.linkml_model import SlotDefinition
+import pytest
 
 from linkml_map.transformer.object_transformer import ObjectTransformer
-from tests.conftest import add_to_integration, setup_integration
-from tests.scaffold import new_scaffold
-
-@pytest.fixture
-def scaffold():
-    """Fresh scaffold per test."""
-    return new_scaffold()
+from tests.conftest import add_to_test_setup, TEST_SETUP_FUNCTIONS
 
 def run_transformer(scaffold):
     """Helper function to run the object transformer with the given scaffold."""
@@ -39,46 +33,44 @@ def test_basic_person_to_agent(scaffold):
     result = run_transformer(scaffold)
     assert result == scaffold["expected"]
 
-@add_to_integration
+# Each setup function should adjust the scaffold (schemas, transform spec, and expected output).
+# Mark with @add_to_test_setup to include in both parameterized unit tests and integration tests.
+#
+# @add_to_test_setup
+# def setup_your_test_name(scaffold):
+#     # Add your test setup code modifying the scaffold here
+
+@add_to_test_setup
 def setup_value_slot_derivation(scaffold):
-    # Add study_name slot to target schema
     add_slot_to_class(scaffold, "Agent", "study_name", "string")
 
-    # Add value slot_derivation
     scaffold["transform_spec"]["class_derivations"]["Agent"]["slot_derivations"]["study_name"] = {
         "value": "Framingham Heart Study",
     }
 
-    # Update expected output
     scaffold["expected"]["study_name"] = "Framingham Heart Study"
 
-def test_value_slot_derivation(scaffold):
-    setup_value_slot_derivation(scaffold)
-
-    result = run_transformer(scaffold)
-    assert result == scaffold["expected"]
-
-@add_to_integration
+@add_to_test_setup
 def setup_value_attribute_slot_derivation(scaffold):
-    # Add study_name slot to target schema
     add_attribute_to_class(scaffold, "Agent", "location", "string")
 
-    # Add value slot_derivation
     scaffold["transform_spec"]["class_derivations"]["Agent"]["slot_derivations"]["location"] = {
         "value": "Framingham",
     }
 
-    # Update expected output
     scaffold["expected"]["location"] = "Framingham"
 
-def test_value_attribute_slot_derivation(scaffold):
-    setup_value_attribute_slot_derivation(scaffold)
-
+@pytest.mark.parametrize(
+    "setup_func",
+    TEST_SETUP_FUNCTIONS,
+    ids=[f.__name__ for f in TEST_SETUP_FUNCTIONS],
+)
+def test_with_setup(scaffold, setup_func):
+    """Apply a setup function, run transformer, and assert expected output."""
+    setup_func(scaffold)
     result = run_transformer(scaffold)
     assert result == scaffold["expected"]
 
-def test_z_integration(scaffold):
-    setup_integration(scaffold)
-
-    result = run_transformer(scaffold)
-    assert result == scaffold["expected"]
+def test_all_setup_integration(integration_scaffold):
+    result = run_transformer(integration_scaffold)
+    assert result == integration_scaffold["expected"]
